Moved reading stdin in parse_pre_push function
git uses the stdin to communicate information about ongoing
actions to the hook. So far the read call for stdin was performed
in init, but not all hooks require it.
It is now moved to the parse_pre_push function which is the only
place where the stdin is used.